changed buffer length type to uint8

+++ rx_uart/driver/uart.c
@@ -192,9 +192,9 @@ uart0_tx_buffer(uint8 *buf, uint16 len)
 }
 
 void ICACHE_FLASH_ATTR
-uart1_tx_buffer(uint8 *buf, uint16 len)
+uart1_tx_buffer(uint8 *buf, uint8 len)
 {
-    uint16 i;
+    uint8 i;
     for (i = 0; i < len; i++)
     {
 	    uart_tx_one_char(UART1, buf[i]);
